feat: show run time & kill count in history modal
