+++ src/render/components/session-timer.jsx
@@ -1,16 +1,6 @@
 import React, { useEffect, useState } from 'react';
 
-const pad = string_ => string_ < 10 ? `0${string_}` : string_;
-
-function formatTime(seconds) {
-  const values = {
-    hours: pad(Math.max(Math.floor(seconds / 3600) % 24, 0)),
-    minutes: pad(Math.max(Math.floor(seconds / 60) % 60, 0)),
-    seconds: pad(Math.max(Math.floor(seconds % 60), 0)),
-  };
-
-  return `${values.hours}:${values.minutes}:${values.seconds}`;
-}
+import { formatTime } from '../utils/formatting';
 
 const defaultTime = '00:00:00';
